requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==4.4.2
filelock==3.20.0
fsspec==2025.9.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
pillow==11.3.0
proglog==0.1.12
PyQt5==5.15.10
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
regex==2025.9.18
requests==2.32.5
setuptools==70.2.0
sympy==1.13.1
tiktoken==0.12.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.15.0
urllib3==2.5.0
